api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
courlan==0.9.3
dateparser==1.1.8
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
htmldate==1.4.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jusText==3.0.0
langcodes==3.3.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.19
tld==0.13
tldextract==3.4.4
tqdm==4.66.0
trafilatura==1.6.1
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
#
faiss-cpu
numpy
